Fix l2 sequencer issues (#2203)
* Revert "Add support for Starkware to layer2 sequencer health EA (#2039)"

This reverts commit e60cb853e7455cc27a6837d984ca50333df5a005.

* Revert
