+++ .yarnrc.yml
@@ -64,4 +64,3 @@ plugins:
 
 pnpEnableEsmLoader: false
 yarnPath: .yarn/releases/yarn-3.2.2.cjs
-
