+++ packages/sources/layer2-sequencer-health/package.json
@@ -30,7 +30,6 @@
     "@chainlink/ea-bootstrap": "workspace:*",
     "@chainlink/ea-test-helpers": "workspace:*",
     "ethers": "^5.4.6",
-    "starknet": "4.1.0",
     "tslib": "^2.3.1"
   },
   "devDependencies": {
