+++ packages/sources/layer2-sequencer-health/schemas/env.json
@@ -49,30 +49,6 @@
       "description": "The blockchain id to connect to",
       "type": "string",
       "default": "1088"
-    },
-    "NUM_RETRIES": {
-      "type": "number",
-      "default": 2
-    },
-    "RETRY_INTERVAL": {
-      "type": "number",
-      "default": 10
-    },
-    "STARKWARE_SEQUENCER_ENDPOINT": {
-      "type": "string",
-      "default": "https://alpha-mainnet.starknet.io"
-    },
-    "STARKWARE_FEEDER_GATEWAY_URL": {
-      "type": "string",
-      "default": "feeder_gateway"
-    },
-    "STARKWARE_GATEWAY_URL": {
-      "type": "string",
-      "default": "gateway"
-    },
-    "STARKWARE_DUMMY_ACCOUNT_ADDRESS": {
-      "type": "string",
-      "default": "0x00000000000000000000000000000000000000000000000000000000000001"
     }
   },
   "allOf": [
