+++ packages/sources/layer2-sequencer-health/src/network.ts
@@ -1,46 +1,24 @@
-import { Logger, Requester, util } from '@chainlink/ea-bootstrap'
-import { HEALTH_ENDPOINTS, Networks, ExtendedConfig } from './config'
+import { Logger, Requester, AxiosRequestConfig } from '@chainlink/ea-bootstrap'
+import { CHAIN_IDS, HEALTH_ENDPOINTS, Networks, RPC_ENDPOINTS } from './config'
+import { BigNumber, ethers } from 'ethers'
+import { AdapterResponseEmptyError } from '@chainlink/ea-bootstrap'
 
-import {
-  checkStarkwareSequencerPendingTransactions,
-  sendDummyStarkwareTransaction,
-} from './starkware'
-import { checkOptimisticRollupBlockHeight, sendEVMDummyTransaction } from './evm'
-
+const DEFAULT_PRIVATE_KEY = '0x0000000000000000000000000000000000000000000000000000000000000001'
 const NO_ISSUE_MSG =
   'This is an error that the EA uses to determine whether or not the L2 Sequencer is healthy.  It does not mean that there is an issue with the EA.'
 
-// These errors come from the Sequencer when submitting an empty transaction
-const sequencerOnlineErrors: Record<Networks, string[]> = {
-  [Networks.Arbitrum]: ['gas price too low', 'forbidden sender address', 'intrinsic gas too low'],
-  // TODO: Optimism error needs to be confirmed by their team
-  [Networks.Optimism]: ['cannot accept 0 gas price transaction'],
-  [Networks.Metis]: ['cannot accept 0 gas price transaction'],
-  // Sending an empty transaction to the dummy Starknet address should return one
-  // of the following error messages.  The Sequencer is considered healthy if the
-  // EA gets back one of the errors below.
-  // UNINITIALIZED_CONTRACT is thrown whenever the dummy address has not been
-  // deployed to the network
-  // OUT_OF_RANGE_FEE is thrown when the network detects a transaction sent
-  // with 0 gas
-  [Networks.Starkware]: [
-    'StarknetErrorCode.UNINITIALIZED_CONTRACT',
-    'StarknetErrorCode.OUT_OF_RANGE_FEE',
-  ],
-}
-
 export interface NetworkHealthCheck {
-  (network: Networks, config: ExtendedConfig): Promise<undefined | boolean>
+  (network: Networks, delta: number, deltaBlocks: number): Promise<undefined | boolean>
 }
 
 export interface ResponseSchema {
   result: number
 }
 
-export const checkSequencerHealth: NetworkHealthCheck = async (
+export const getSequencerHealth: NetworkHealthCheck = async (
   network: Networks,
 ): Promise<undefined | boolean> => {
-  if (!HEALTH_ENDPOINTS[network]?.endpoint) {
+  if (!HEALTH_ENDPOINTS[network].endpoint) {
     Logger.info(`Health endpoint not available for network: ${network}`)
     return
   }
@@ -56,99 +34,121 @@ export const checkSequencerHealth: NetworkHealthCheck = async (
   return isHealthy
 }
 
+export const requestBlockHeight = async (network: Networks): Promise<number> => {
+  const request: AxiosRequestConfig = {
+    method: 'POST',
+    url: RPC_ENDPOINTS[network],
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    data: {
+      jsonrpc: '2.0',
+      method: 'eth_blockNumber',
+      params: [],
+      id: 1,
+    },
+  }
+  const response = await Requester.request<ResponseSchema>(request)
+  const hexBlock = response.data.result
+  if (!hexBlock) {
+    throw new AdapterResponseEmptyError({
+      message: `Block number not found on network: ${network}`,
+    })
+  }
+  return BigNumber.from(hexBlock).toNumber()
+}
+
+// TODO: Implement when ready
+export const getL1RollupStatus: NetworkHealthCheck = async (): Promise<boolean> => {
+  return true
+}
+
 export const getStatusByTransaction = async (
   network: Networks,
-  config: ExtendedConfig,
+  timeout: number,
 ): Promise<boolean> => {
-  let isSequencerHealthy = true
-  try {
-    Logger.info(`Submitting empty transaction for network: ${network}`)
-    await sendEmptyTransaction(network, config)
-  } catch (e) {
-    isSequencerHealthy = isExpectedErrorMessage(network, e as Error)
-  }
-  return isSequencerHealthy
-}
+  const rpcEndpoint = RPC_ENDPOINTS[network]
+  const chainId = CHAIN_IDS[network]
+  const provider = new ethers.providers.JsonRpcProvider(rpcEndpoint, chainId)
+  const wallet = new ethers.Wallet(DEFAULT_PRIVATE_KEY, provider)
 
-const sendEmptyTransaction = async (network: Networks, config: ExtendedConfig): Promise<void> => {
-  switch (network) {
-    case Networks.Starkware:
-      await sendDummyStarkwareTransaction(config)
-      break
-    default:
-      await sendEVMDummyTransaction(network, config.timeoutLimit)
+  // These errors come from the Sequencer when submitting an empty transaction
+  const sequencerOnlineErrors: Record<Networks, string[]> = {
+    [Networks.Arbitrum]: ['gas price too low', 'forbidden sender address', 'intrinsic gas too low'],
+    // TODO: Optimism error needs to be confirmed by their team
+    [Networks.Optimism]: ['cannot accept 0 gas price transaction'],
+    [Networks.Metis]: ['cannot accept 0 gas price transaction'],
   }
-}
 
-const isExpectedErrorMessage = (network: Networks, e: Error) => {
+  const networkTx: Record<Networks, ethers.providers.TransactionRequest> = {
+    // Arbitrum zero gas price will be auto adjusted by the network to the minimum
+    [Networks.Arbitrum]: {
+      value: 0,
+      gasLimit: 0,
+      gasPrice: 1,
+      to: wallet.address,
+    },
+    [Networks.Optimism]: {
+      value: 0,
+      gasLimit: 0,
+      gasPrice: 0,
+      to: wallet.address,
+    },
+    [Networks.Metis]: {
+      value: 0,
+      gasLimit: 0,
+      gasPrice: 0,
+      to: wallet.address,
+    },
+  }
   const _getErrorMessage = (e: Error): string => {
     const paths = {
       [Networks.Arbitrum]: ['error', 'message'],
       [Networks.Optimism]: ['error', 'message'],
       [Networks.Metis]: ['error', 'message'],
-      [Networks.Starkware]: ['errorCode'],
     }
     return (Requester.getResult(e, paths[network]) as string) || ''
   }
-  const error = e as Error
-  if (sequencerOnlineErrors[network].includes(_getErrorMessage(error))) {
-    Logger.debug(`Transaction submission failed with an expected error ${_getErrorMessage(error)}.`)
-    return true
-  }
-  Logger.error(
-    `Transaction submission failed with an unexpected error. ${NO_ISSUE_MSG} Error Message: ${error.message}`,
-  )
-  return false
-}
 
-export const checkNetworkProgress: NetworkHealthCheck = (
-  network: Networks,
-  config: ExtendedConfig,
-) => {
-  switch (network) {
-    case Networks.Starkware:
-      return checkStarkwareSequencerPendingTransactions()(config)
-    default:
-      return checkOptimisticRollupBlockHeight(network)(config)
-  }
-}
-
-export async function retry<T>({
-  promise,
-  retryConfig,
-}: {
-  promise: () => Promise<T>
-  retryConfig: ExtendedConfig['retryConfig']
-}): Promise<T> {
-  let numTries = 0
-  let error
-  while (numTries < retryConfig.numRetries) {
-    try {
-      return await promise()
-    } catch (e) {
-      error = e
-      numTries++
-      await util.sleep(retryConfig.retryInterval)
+  try {
+    Logger.info(`Submitting empty transaction for network: ${network}`)
+    const receipt = await race({
+      timeout,
+      promise: wallet.sendTransaction(networkTx[network]),
+      error: `Transaction receipt not received in ${timeout} milliseconds`,
+    })
+    Logger.info(`Transaction receipt received with hash ${receipt.hash} for network: ${network}`)
+    return (await receipt.wait()).confirmations > 0
+  } catch (e: any) {
+    const error = e as Error
+    if (sequencerOnlineErrors[network].includes(_getErrorMessage(error))) {
+      Logger.debug(
+        `Transaction submission failed with an expected error ${_getErrorMessage(error)}.`,
+      )
+      return true
     }
+    Logger.error(
+      `Transaction submission failed with an unexpected error. ${NO_ISSUE_MSG} Error Message: ${error.message}`,
+    )
+    return false
   }
-  throw error
 }
 
-export function race<T>({
+export function race({
   promise,
   timeout,
   error,
 }: {
-  promise: Promise<T>
+  promise: Promise<ethers.providers.TransactionResponse>
   timeout: number
   error: string
-}): Promise<T> {
+}): Promise<ethers.providers.TransactionResponse> {
   let timer: NodeJS.Timeout
 
   return Promise.race([
     new Promise((_, reject) => {
       timer = setTimeout(reject, timeout, error)
-    }) as Promise<T>,
+    }) as Promise<ethers.providers.TransactionResponse>,
     promise.then((value) => {
       clearTimeout(timer)
       return value
